move flushop to db dir
